Rough of Talks page blog.
Yikes! I'm over a day late on this one. But the talks page must go on.
I'm happier with this blog that some of the others. I think it got me
thinking about short vs long term content.
"I made a Talks Page in Jekyll" is only interesting for the stream, but
"Here's how to make nice Twitter cards with examples in Jekyll" has more
long-term usefulness. Thinking about things in terms of "Would Mia be
able to implement this after reading this blog?" helped drive it, but
ultimately, I don't think this delivered on that due to deadlines.